% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_parsers.R
\name{annual_totals_parser}
\alias{annual_totals_parser}
\title{Annual total parser}
\usage{
annual_totals_parser(response)
}
\arguments{
\item{response}{a httr response}
}
\value{
parsed response from httr
}
\description{
Annual total parser
}
