% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{webtris_areas}
\alias{webtris_areas}
\title{Areas}
\usage{
webtris_areas(areaID)
}
\arguments{
\item{areaID}{area ID. If none provided then returns all.}
}
\value{
data.frame
}
\description{
Retruns areas from WEBTRIS API
}
\examples{
webtris_areas()
webtris_areas(areaID=3)
}
