% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tidy.R
\name{webtris_report}
\alias{webtris_report}
\title{webtris_report}
\usage{
webtris_report(sites, start_date, end_date, report_type, maxrequest = 10000,
  verbose = FALSE, return_raw = FALSE)
}
\arguments{
\item{sites}{siteId(s) i.e. 7 or '7' or c('7','8', '9')}

\item{start_date}{in format 'DDMMYYYY' or as a Date object}

\item{end_date}{in format 'DDMMYYYY' or as a Date object}

\item{report_type}{options are: "daily", "annual-totals", "annual-monthly", "monthly-summary", "monthly-daily", "monthly-daily-aggregate", "monthly-hourly-aggregate"}

\item{maxrequest}{Maximum number of rows to return per requiest (optional)}

\item{verbose}{Prints the API call}

\item{return_raw}{Return a raw httr response rather than a data frame}
}
\value{
a dataframe (default) or httr object (if return_raw=TRUE)
}
\description{
webtris_report
}
\examples{
webtris_report(sites=c("7"), start_date="01-01-2017", end_date="01-03-2017", report_type ='daily')
}
