% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{webtris_sites}
\alias{webtris_sites}
\title{Sites}
\usage{
webtris_sites(siteID, sf = FALSE, crs = 4326)
}
\arguments{
\item{siteID}{Numeric site ID. Blank will return all sites.}

\item{sf}{logic, return results as an sf object, defaults `FALSE`.}

\item{crs}{Numeric, specify CRS, defaults to 4326}
}
\value{
dataframe of sites.
}
\description{
Sites
}
\examples{
webtris_sites()
webtris_sites('3')
webtris_sites('3', sf = TRUE)
}
