% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{webtris_quality}
\alias{webtris_quality}
\title{Quality}
\usage{
webtris_quality(siteID, start_date, end_date, daily = TRUE)
}
\arguments{
\item{siteID}{either a single site ID or a list c(1,2,3). Daily data doesn't support a list of sites.}

\item{start_date}{either a string in the format ddmmyy or a Date (lubridate)}

\item{end_date}{either a string in the format ddmmyy or a Date (lubridate)}

\item{daily}{TRUE or FALSE. If FALSE returns overall. Use FALSE when list of siteIDs.}
}
\value{
data.frame
}
\description{
Quality
}
\note{
The API returns incorrect quality values for aggregated queries, fix is implemented on the client side.
}
\examples{
webtris_quality('7', start_date='01012017', end_date='01022017', daily=TRUE)
}
