% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_wrappers.R
\name{webtris_sites}
\alias{webtris_sites}
\title{Sites}
\usage{
webtris_sites(siteID)
}
\arguments{
\item{siteID}{Numeric site ID. Blank will return all sites.}
}
\value{
dataframe of sites.
}
\description{
Sites
}
\examples{
webtris_sites()
webtris_sites('3')
}
